+++ package/src/extensions/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "async_std")]
+pub mod async_std;
+
+#[cfg(feature = "tokio")]
+pub mod tokio;