+++ package/src/extensions/tokio/mod.rs
@@ -1,30 +1,27 @@
 use std::path::{Path, PathBuf};
 
-use tokio::{
-    fs::{self, ReadDir},
-    io,
-};
+use tokio::{fs, io};
 
-use crate::{GetDir, Target};
+use crate::{structs::target::Target, util::GetDir};
 
-async fn target_exists<'a>(
+async fn is_target_exists<'a>(
     path: &'a Path,
-    target: &'a Target<'a>,
+    target: &'a Target,
 ) -> bool {
     match target {
         | Target::Dir(tg) => {
-            let target_path: PathBuf = path.join(tg.name);
+            let target_path: PathBuf = path.join(tg.name.clone());
 
-            if target_path.exists() && target_path.is_dir() {
+            if target_path.is_dir() {
                 return true;
             }
 
             false
         },
         | Target::File(tg) => {
-            let target_path: PathBuf = path.join(tg.name);
+            let target_path: PathBuf = path.join(tg.name.clone());
 
-            if target_path.exists() && target_path.is_file() {
+            if target_path.is_file() {
                 return true;
             }
 
@@ -35,10 +32,10 @@ async fn target_exists<'a>(
 
 async fn search_targets<'a>(
     dir: &'a PathBuf,
-    targets: &'a Vec<Target<'a>>,
+    targets: &'a Vec<Target>,
 ) -> Option<PathBuf> {
     for target in targets {
-        if target_exists(dir, target).await {
+        if is_target_exists(dir, target).await {
             return Some(dir.to_owned());
         }
     }
@@ -46,15 +43,15 @@ async fn search_targets<'a>(
     None
 }
 
-async fn search_dir<'a>(
-    dir: &'a PathBuf,
-    targets: &'a Vec<Target<'a>>,
+async fn get_dir(
+    dir: &PathBuf,
+    targets: &Vec<Target>,
 ) -> io::Result<PathBuf> {
     if let Some(found) = search_targets(dir, targets).await {
         return Ok(found);
     }
 
-    let mut entries: ReadDir = fs::read_dir(dir).await?;
+    let mut entries: fs::ReadDir = fs::read_dir(dir).await?;
 
     while let Some(entry) = entries.next_entry().await? {
         let current: PathBuf = entry.path();
@@ -64,8 +61,23 @@ async fn search_dir<'a>(
                 return Ok(found);
             }
 
-            if let Ok(found) = Box::pin(search_dir(&current, targets)).await {
+            if let Ok(found) = Box::pin(get_dir(&current, targets)).await {
                 return Ok(found);
+            }
+        }
+    }
+
+    Err(io::Error::from(io::ErrorKind::NotFound))
+}
+
+async fn get_dir_reverse(
+    dir: &Path,
+    targets: &Vec<Target>,
+) -> io::Result<PathBuf> {
+    for ancestor in dir.ancestors() {
+        for target in targets {
+            if is_target_exists(ancestor, target).await {
+                return Ok(ancestor.to_path_buf());
             }
         }
     }
@@ -86,20 +98,12 @@ pub trait GetDirAsyncExt {
     ) -> impl std::future::Future<Output = io::Result<PathBuf>> + Send;
 }
 
-impl GetDirAsyncExt for GetDir<'_> {
+impl GetDirAsyncExt for GetDir {
     async fn run_async(&self) -> io::Result<PathBuf> {
-        search_dir(&self.dir, &self.targets).await
+        get_dir(&self.dir, &self.targets).await
     }
 
     async fn run_reverse_async(&self) -> io::Result<PathBuf> {
-        for ancestor in self.dir.ancestors() {
-            for target in &self.targets {
-                if target_exists(ancestor, target).await {
-                    return Ok(ancestor.to_path_buf());
-                }
-            }
-        }
-
-        Err(io::Error::from(io::ErrorKind::NotFound))
+        get_dir_reverse(&self.dir, &self.targets).await
     }
 }