+++ tests/src/async_std.rs
@@ -12,7 +12,7 @@ mod tests {
     #[async_std::test]
     async fn test_get_dir_by_target_dir() {
         let dir: PathBuf = GetDir::new()
-            .targets(vec![Target::Dir(DirTarget { name: "src" })])
+            .target(Target::Dir(DirTarget::new("src")))
             .run_async()
             .await
             .unwrap();
@@ -26,7 +26,7 @@ mod tests {
     #[async_std::test]
     async fn test_get_dir_by_target_file() {
         let dir: PathBuf = GetDir::new()
-            .targets(vec![Target::File(FileTarget { name: "Cargo.toml" })])
+            .target(Target::File(FileTarget::new("Cargo.toml")))
             .run_async()
             .await
             .unwrap();
@@ -40,7 +40,7 @@ mod tests {
     #[async_std::test]
     async fn test_get_dir_by_tarrun_reverse_dir() {
         let dir: PathBuf = GetDir::new()
-            .targets(vec![Target::Dir(DirTarget { name: "target" })])
+            .target(Target::Dir(DirTarget::new("target")))
             .run_reverse_async()
             .await
             .unwrap();
@@ -54,7 +54,7 @@ mod tests {
     #[async_std::test]
     async fn test_get_dir_by_tarrun_reverse_file() {
         let dir: PathBuf = GetDir::new()
-            .targets(vec![Target::File(FileTarget { name: "LICENSE" })])
+            .target(Target::File(FileTarget::new("LICENSE")))
             .run_reverse_async()
             .await
             .unwrap();
@@ -69,7 +69,7 @@ mod tests {
     async fn test_get_dir_by_target_file_in_specific_dir() {
         let dir: PathBuf = GetDir::new()
             .directory(current_dir().unwrap().join("..").join("package"))
-            .targets(vec![Target::File(FileTarget { name: "lib.rs" })])
+            .target(Target::File(FileTarget::new("lib.rs")))
             .run_async()
             .await
             .unwrap();